Save creation requests to database
Adding Mongoose, which means I'm not sure if I actually need to keep the
mongodb instance in app.locals as well.

I needed to constrain mongodb because connect-mongo expects ^4.1.0, but
Mongoose expects exactly 4.1.4. NPM was installling two versions, and
that was causing issues with Typescript because mongodb 4.2.0 adds the
"srvMaxHosts" and "srvServiceName" options to the MongoClient, which
meant that Typescript saw mongoose MongoClients and mongodb MongoClients
as incompatible.