Start implementing account creation
We outsource local user operations to some yet-to-be-defined other
process.

This will be better because we can have a thin server running as root
on the SSH server or whatever, and communicate over an API. Then we can
e.g. have SAUCE in a docker container or something without any issue.