Make password reset better on slow connections
Now we load the big zxcvbn dictionaries asynchronously, which means that
our time-to-full-paint is much faster now. Also introduces some logic to
wait for the scripts to load and check password strength before
submitting the form—previously on a slow connection it was possible to
submit a bad password if you were quick about it, which the server would
validate and give you a nasty 400 error page.